Adapt to go 1.13 module support
The module support in go 1.13 has become quite awesome with the
introduction of module proxies which greatly speed up fetching of
dependencies, and also provides security with regards to tampering and
missing packages.

This PR adapts the project to rely solely on go modules without vendored
packages.

Changes:
- Make travis use go 1.13
- Make dockerfile build the project using a multistage build
- Add Makefile for build commands and tooling
- Remove Gopkg files and vendor directory
- Add go module file and sum file
- Add tools.go to include used tools in the module
- Adapt crd and rds provider to upgraded modules

We might want travis to simply run docker build and push.