Name Subnet Groups after the VpcId
Currently, the SubnetGroup get it's name from the database name. This
leads to the creation of groups composed by the same subnets with
different names which consume all the Subnet Group quota from an AWS
account. Since the list of subnets comes from the VPC of one of the K8s
nodes, naming the groups after the VPC Id generate unique Subnet Groups
minimizing quota problem.