POPS-1461 Disable delete function
We are about to migrate services to EKS which involves deploying helm charts to EKS and later removing the charts
from the legacy cluster. It can happen that when charts are removed from the legacy cluster, this operator would
delete a RDS instance in AWS (because the the chart does not have db.Spec.DeleteProtection). Therefore, to completely
eliminate any possibilities of this happening, this delete function does not do anything.